bots/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
atproto==0.0.48
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
dnspython==2.6.1
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.26.0
idna==3.7
libipld==1.2.3
pillow==10.3.0
pycparser==2.22
pydantic==2.7.3
pydantic_core==2.18.4
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.2.1
websockets==12.0
